tmp fix

+++ src/main/resources/db/migration/V1__CREATE_USER.sql
@@ -1,7 +1,7 @@
 create table "user" (
     id serial constraint user_pk primary key,
     email varchar(255),
-    password integer,
+    password bigint,
     file varchar(255),
     theme varchar(255),
     font varchar(50),
